Bugfix: Xcodeproj wasn't being installed.
Simply declaring Xcodeproj as a resource doesn't actually install it, it just downloads the necessary files.

The Xcodeproj installation can't happen normally because it needs sudo; instead, we need to install it to Gryphon's homebrew folder, and tell Gryphon where that folder is.

This commit makes the formula install Xcodeproj, then alter one of Gryphon's Swift files so that the new executable will generate a bash script that points to the correct installation folder.